refactor(template): move ParseApp method to separate file
