+++ internal/pkg/template/template.go
@@ -97,29 +97,6 @@ func (t *Template) Parse(path string, data interface{}, options ...ParseOption) 
 	return &Content{buf}, nil
 }
 
-// ParseApp parses an application's CloudFormation template with the specified data object and returns its content.
-func (t *Template) ParseApp(name string, data interface{}, options ...ParseOption) (*Content, error) {
-	tpl, err := t.parse("base", fmt.Sprintf(fmtAppCFTemplatePath, name), options...)
-	if err != nil {
-		return nil, err
-	}
-	for _, templateName := range commonCFTemplateNames {
-		nestedTpl, err := t.parse(templateName, fmt.Sprintf(fmtCommonCFTemplatePath, templateName), options...)
-		if err != nil {
-			return nil, err
-		}
-		_, err = tpl.AddParseTree(templateName, nestedTpl.Tree)
-		if err != nil {
-			return nil, fmt.Errorf("add parse tree of %s to base template: %w", templateName, err)
-		}
-	}
-	buf := &bytes.Buffer{}
-	if err := tpl.Execute(buf, data); err != nil {
-		return nil, fmt.Errorf("execute app template %s with data %v: %w", name, data, err)
-	}
-	return &Content{buf}, nil
-}
-
 // WithFuncs returns a template that can parse additional custom functions.
 func WithFuncs(fns map[string]interface{}) ParseOption {
 	return func(t *template.Template) *template.Template {
